Add displayController variables and event listeners
Only the name change event listener has been fully
implemented; the other functions will be completed
in future commits, but I wanted to write them in so
they aren't forgotten.
The Player class was also updated with a third score
parameter to allow scorekeeping. Both the Player
and gameController classes got a couple new getters
and setters, but nothing major.